// packages/curve-ui-kit/src/themes/model/figma-tokens.ts
export const FIGMA_TOKENS = {
// ...
  primitives: {
    grays: {
      '10': '#fdfcfcff',
      '25': '#fafafaff',
      '50': '#f8f7f7ff',
      '75': '#f6f4f4ff',
      '100': '#eeecebff',
      '150': '#e7e4e2ff',
      '200': '#dedbd8ff',
      '300': '#d4d0ccff',
      '400': '#bbb6afff',
      '500': '#968e84ff',
      '600': '#746e66ff',
      '700': '#5a554fff',
      '750': '#494540ff',
      '800': '#3b3834ff',
      '850': '#302e2aff',
      '900': '#252420ff',
      '950': '#1f1e1bff',
      '975': '#191815ff',
    },
// ...
  },
  effects: {
    activetabshadow: {
      '0': {
        shadowType: 'dropShadow',
        radius: 0,
        color: '#0e3f7e0a',
        offsetX: 0,
        offsetY: 0,
        spread: 1,
      },
      '1': {
        shadowType: 'dropShadow',
        radius: 1,
        color: '#2a33450a',
        offsetX: 1,
        offsetY: 0,
        spread: -0.5,
      },
      '2': {
        shadowType: 'dropShadow',
        radius: 3,
        color: '#2a33460a',
        offsetX: 3,
        offsetY: 0,
        spread: -1.5,
      },
// ...
    },
// ...
  },
// ...
} as const